pyproject.toml
[tool.poetry]
name = "tidb-evaluation"
version = "0.1.0"
description = ""
authors = ["Cheese <cheese.javaee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
logging = "^0.4.9.6"
pydantic = "^2.9.2"
httpx = "^0.27.2"
click = "^8.1.7"
retry = "^0.9.2"
python-dotenv = "^1.0.1"
openai = "^1.55.0"
pandas = "^2.2.3"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
